# openapi.yaml
openapi: 3.0.3
info:
  title: Employee Attendance & Analytics API
  version: 2.0.0
  description: |
    Contract for the HROne engineering assignment. Your FastAPI app must follow this document exactly: paths, methods,
    parameter names, status codes and response field names. The application is organized into routers, schemas, services, configuration, and database modules;
    `app/main.py` exposes the FastAPI entry point and must start with `uvicorn app.main:app --port 8000`.

    **Where things are defined.** This file defines the HTTP contract and the business rules R1-R10. What is stored in
    MongoDB (collections, field types, what the grader seeds, nine sample documents) is defined in `DATA_MODEL.md`.

    ## Configuration
    `MONGO_URI` and `MONGO_DB` are read from environment variables (a local `.env` is fine; real environment variables
    must win). Create all indexes at startup, idempotently.

    ## Conventions used by every endpoint
    * **Employees are identified by `emp_code`.** It is a customer-generated business code assigned by the customer's HR
      team (this API does not generate it). It is unique, never changes, and is meant for humans. It is **not** MongoDB's
      `_id`. `_id` is internal: it never appears in a request or a response. Attendance records have no id either; a
      record is addressed by (`emp_code`, `date`).
    * **Instants are epoch milliseconds.** Every point in time in a request or response (`punched_at`, `punch_in`,
...

...

...

...

...

...

...

...
paths:
...
  /employees:
    post:
      tags: [employees]
      operationId: createEmployee
      summary: Create an employee
      description: |
        The client supplies `emp_code` (the customer's own code). It must be unique, **enforced by a database unique
        index** (concurrent duplicate creates must yield exactly one 201 and the rest 409). `shift_start` must differ
        from `shift_end`. `created_at` is set by the server. There is no endpoint to change an `emp_code`.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/EmployeeCreate'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Employee'}
        '409': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/ValidationError'}
    get:
      tags: [employees]
      operationId: listEmployees
      summary: List employees
      description: Sorted by `emp_code` ascending. `total` counts the filtered set, not the whole collection.
      parameters:
        - name: department
          in: query
          required: false
          schema: {type: string}
          description: Exact, case-sensitive match.
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
      responses:
        '200':
          description: A page of employees
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EmployeePage'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /attendance/punch-in:
    post:
      tags: [attendance]
      operationId: punchIn
      summary: Record a punch-in
      description: |
        Creates the day's record. `date` follows R1 and `late_minutes` follows R2; `punch_out` and `work_hours` start as
        `null`, `overtime_minutes` as 0, `half_day` as false, `history` as `[]`. If `punched_at` is omitted, use the
        current time. `status` defaults to `PRESENT`; only `PRESENT`, `WFH` and `ON_DUTY` are accepted here.
        **Concurrency:** if the same employee punches in twice for the same attendance date at the same moment,
        exactly one request gets 201 and the rest get 409. Do not rely on check-then-insert.
        Errors: 404 unknown employee; 409 already punched in for that date; 422 invalid body.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PunchInRequest'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AttendanceRecord'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /attendance/punch-out:
    post:
      tags: [attendance]
      operationId: punchOut
      summary: Record a punch-out
      description: |
        Closes the employee's most recent record whose `punch_in` is at or before `punched_at` (this must work across
        midnight for overnight shifts). Computes `work_hours` (R4), `overtime_minutes` (R3) and `half_day` (R5).
        Errors: 404 unknown employee, or no punch-in found; 409 that record is already punched out (two simultaneous
        punch-outs must yield exactly one 200); 422 if `punched_at` <= `punch_in` or more than 24 hours after it.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PunchOutRequest'}
      responses:
        '200':
          description: Updated record
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AttendanceRecord'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /attendance:
    get:
      tags: [attendance]
      operationId: listAttendance
      summary: List attendance records
      description: |
        Sorted by `date` descending, then `emp_code` ascending. Must stay fast on a collection of ~100,000 records;
        do not load the collection into memory. `date_from`/`date_to` are inclusive; `date_from` > `date_to` -> 422.
      parameters:
        - {name: emp_code, in: query, required: false, schema: {type: string}}
        - {name: date_from, in: query, required: false, schema: {type: string, format: date}}
        - {name: date_to, in: query, required: false, schema: {type: string, format: date}}
        - {name: status, in: query, required: false, schema: {$ref: '#/components/schemas/Status'}}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
      responses:
        '200':
          description: A page of records
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AttendancePage'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /attendance/{emp_code}/{date}:
    patch:
      tags: [attendance]
      operationId: regularizeAttendance
      summary: Regularize (manually correct) one day's record
      description: |
        The record is addressed by its natural key: the employee's `emp_code` and the attendance `date`
        (`YYYY-MM-DD`, the record's `date` field).
        Omitted fields are unchanged. After applying the change, **recompute** `late_minutes`, `work_hours`,
        `overtime_minutes` and `half_day` from the final times (R2-R5) and **append** one entry to `history` (never
        overwrite it); see "History" above. `history[].changes` contains only the fields that actually changed.
        Rules: setting status `ABSENT` or `LEAVE` clears both punch times, so supplying `punch_in` or `punch_out` with it
        is 422; a presence status requires a `punch_in` (supplied or already stored); `punch_in` must stay on the record's
        attendance date (R1); `punch_out` must be after `punch_in` and within 24 hours; a request that changes nothing
        is 422. 404 if the employee does not exist or has no record for that date. Concurrent edits must not silently
        lose each other's history (409 is acceptable for the loser).
      parameters:
        - {name: emp_code, in: path, required: true, schema: {type: string}}
        - {name: date, in: path, required: true, schema: {type: string, format: date}, description: "The record's attendance date, YYYY-MM-DD."}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RegularizeRequest'}
      responses:
        '200':
          description: Updated record
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AttendanceRecord'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/ValidationError'}
...
components:
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema: {type: integer, minimum: 1, default: 1}
    PageSize:
      name: page_size
      in: query
      required: false
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            type: object
            required: [detail]
            properties:
              detail: {type: string}
    ValidationError:
      description: Validation error (FastAPI default body)
      content:
        application/json:
          schema:
            type: object
            required: [detail]
            properties:
              detail: {type: array, items: {type: object, additionalProperties: true}}
  schemas:
    EpochMillis:
      type: integer
      format: int64
      minimum: 100000000000
      maximum: 4102444800000
      description: |
        An instant as milliseconds since 1970-01-01T00:00:00Z. Seconds-looking values (below 100000000000), floats and
        strings are rejected with 422.
      example: 1783312500000
    EpochMillisNullable:
      type: integer
      format: int64
      nullable: true
      minimum: 100000000000
      maximum: 4102444800000
      description: An instant as epoch milliseconds, or null.
      example: 1783312500000
    Status:
      type: string
      enum: [PRESENT, ABSENT, LEAVE, WFH, ON_DUTY]
      description: See the status table in the API description.
    PresenceStatus:
      type: string
      enum: [PRESENT, WFH, ON_DUTY]
    EmployeeCreate:
      type: object
      required: [emp_code, name, email, department, joined_on]
      properties:
        emp_code:
          type: string
          pattern: '^EMP\d{4,6}$'
          example: EMP0001
          description: |
            The customer's own employee code, supplied by the client. Unique and immutable. Not MongoDB's `_id`.
            Format for this assignment: `EMP` followed by 4-6 digits.
        name: {type: string, minLength: 1, maxLength: 100}
        email: {type: string, maxLength: 120, pattern: '^[^@\s]+@[^@\s]+\.[^@\s]+$'}
        department: {type: string, minLength: 1, maxLength: 50}
        shift_start: {type: string, pattern: '^([01]\d|2[0-3]):[0-5]\d$', default: '09:30', description: "24-hour HH:MM, IST."}
        shift_end: {type: string, pattern: '^([01]\d|2[0-3]):[0-5]\d$', default: '18:30', description: "24-hour HH:MM, IST. <= shift_start means an overnight shift."}
        joined_on: {type: string, format: date, description: "First day of employment, YYYY-MM-DD."}
    Employee:
      type: object
      required: [emp_code, name, email, department, shift_start, shift_end, joined_on, created_at]
      properties:
        emp_code: {type: string, description: "The customer-generated employee code. Unique, immutable. Not the MongoDB _id."}
        name: {type: string}
        email: {type: string}
        department: {type: string}
        shift_start: {type: string}
        shift_end: {type: string}
        joined_on: {type: string, format: date}
        created_at:
          readOnly: true
          allOf:
            - {$ref: '#/components/schemas/EpochMillis'}
          description: Server-set creation time, epoch milliseconds.
    EmployeePage:
      type: object
      required: [items, total, page, page_size]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Employee'}}
        total: {type: integer}
        page: {type: integer}
        page_size: {type: integer}
    PunchInRequest:
      type: object
      required: [emp_code]
      properties:
        emp_code: {type: string}
        punched_at:
          allOf:
            - {$ref: '#/components/schemas/EpochMillis'}
          description: Epoch milliseconds. Defaults to now.
        status:
          allOf:
            - {$ref: '#/components/schemas/PresenceStatus'}
          default: PRESENT
    PunchOutRequest:
      type: object
      required: [emp_code]
      properties:
        emp_code: {type: string}
        punched_at:
          allOf:
            - {$ref: '#/components/schemas/EpochMillis'}
          description: Epoch milliseconds. Defaults to now.
    RegularizeRequest:
      type: object
      required: [reason, regularized_by]
      description: Only these fields can be changed. Derived fields can not be sent.
      properties:
        status: {$ref: '#/components/schemas/Status'}
        punch_in: {$ref: '#/components/schemas/EpochMillis'}
        punch_out: {$ref: '#/components/schemas/EpochMillis'}
        reason: {type: string, minLength: 5, maxLength: 200}
        regularized_by: {type: string, minLength: 1, maxLength: 50}
    HistoryEntry:
      type: object
      required: [at, by, reason, changes]
      properties:
        at: {$ref: '#/components/schemas/EpochMillis'}
        by: {type: string}
        reason: {type: string}
        changes:
          type: object
          description: |
            One key per field that actually changed (`status`, `punch_in`, `punch_out`, `work_hours`, `late_minutes`,
            `overtime_minutes`, `half_day`). The `from` / `to` values of `punch_in` and `punch_out` are epoch
            milliseconds (or null); the others keep their normal types.
          additionalProperties:
            type: object
            properties:
              from: {nullable: true}
              to: {nullable: true}
          example:
            punch_in: {from: 1783312500000, to: 1783310280000}
            late_minutes: {from: 35, to: 0}
    AttendanceRecord:
      type: object
      required: [emp_code, date, status, punch_in, punch_out, work_hours, late_minutes, overtime_minutes, half_day, history]
      properties:
        emp_code: {type: string}
        date: {type: string, format: date, description: "The attendance day (R1), YYYY-MM-DD."}
        status: {$ref: '#/components/schemas/Status'}
        punch_in: {$ref: '#/components/schemas/EpochMillisNullable'}
        punch_out: {$ref: '#/components/schemas/EpochMillisNullable'}
        work_hours: {type: number, nullable: true, readOnly: true, description: Computed (R4). Null until punch-out.}
        late_minutes: {type: integer, readOnly: true, description: Computed (R2).}
        overtime_minutes: {type: integer, readOnly: true, description: Computed (R3).}
        half_day: {type: boolean, readOnly: true, description: Computed (R5).}
        history: {type: array, items: {$ref: '#/components/schemas/HistoryEntry'}}
    AttendancePage:
      type: object
      required: [items, total, page, page_size]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/AttendanceRecord'}}
        total: {type: integer}
        page: {type: integer}
        page_size: {type: integer}
...